+++ MCIA/MCIA/include/MovieService.h
@@ -7,12 +7,12 @@
 
 class MovieService {
 public:	
-	DBPage<WatchedMovieDisplayer> GetWatchedMoviesOfUser(const uint32_t userId, const int page, const int nmbRowsPerPage);
-	DBPage<WishlistedMovieDisplayer> GetWishListOfUser(const uint32_t userId, const int page, const int nmbRowsPerPage);
+	DBPage<WatchedMovieDisplayer> GetWatchedMoviesOfUser(const uint16_t userId, const int page, const int nmbRowsPerPage);
+	DBPage<WishlistedMovieDisplayer> GetWishListOfUser(const uint16_t userId, const int page, const int nmbRowsPerPage);
 	MovieInformationDisplayer GetMovieInformations(const uint32_t id);
-	void AddMovieToWatchlist(const uint32_t userId, const uint32_t movieId, const float rating);
-	void RemoveMovieFromWatchlist(const uint32_t userId, const uint32_t movieId);
-	void AddMovieToWishlist(const uint32_t userId, const uint32_t movieId);
-	void RemoveMovieFromWishlist(const uint32_t userId, const uint32_t movieId);
-	void MoveMovieFromWishlistToWatched(const uint32_t userId, const uint32_t movieId, const float rating);
+	void AddMovieToWatchlist(const uint16_t userId, const uint32_t movieId, const float rating);
+	void RemoveMovieFromWatchlist(const uint16_t userId, const uint32_t movieId);
+	void AddMovieToWishlist(const uint16_t userId, const uint32_t movieId);
+	void RemoveMovieFromWishlist(const uint16_t userId, const uint32_t movieId);
+	void MoveMovieFromWishlistToWatched(const uint16_t userId, const uint32_t movieId, const float rating);
 };
